refactor(harness): move cli package listing projection
